Separate accepted and rejected files
Currently it's not possible to distinguish between accepted and rejected files. The `onDrop` and the `onDropAccepted` functions gets only called with the accepted files. The `onDropRejected` function on the other side gets only called with the rejected files. But sometimes it's useful to react to both states.

This pull request implements the feature in a backward compatible way. The `onDrop` function gets the rejected files passed as second parameter. The `onDropAccepted` and the `onDropRejected` functions behaves the same.